# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedWatch AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Signal {
  id: number;
  created_at: string;
  incident_date?: string | null;
  source_type?: string | null;
  location_district?: string | null;
  location_type?: string | null;
  raw_text?: string | null;
  reporter_type?: string | null;
  drug_name?: string | null;
  hospital_name?: string | null;
  risk_level?: string | null;
  category?: string | null;
  extracted_entities?: string | null;
  recommended_action?: string | null;
  status?: string | null;
  nlp_summary?: string | null;
}

export interface SignalsListResponse {
  signals: Signal[];
  total: number;
}

export type SignalSummaryByRisk = {
  critical: number;
  high: number;
  medium: number;
  low: number;
};

export type SignalSummaryByCategory = {
  adr: number;
  hospital_issue: number;
  outbreak_signal: number;
  treatment_complication: number;
};

export type SignalSummaryByLocationType = {
  urban: number;
  rural: number;
};

export interface SignalSummary {
  total_today: number;
  critical_count: number;
  districts_affected: number;
  signals_last_hour: number;
  by_risk: SignalSummaryByRisk;
  by_category: SignalSummaryByCategory;
  by_location_type: SignalSummaryByLocationType;
}

export interface AnalyzeSignalBody {
  raw_text: string;
  source_type?: string;
  location_district?: string;
  location_type?: string;
  reporter_type?: string;
  drug_name?: string | null;
  hospital_name?: string | null;
  incident_date?: string | null;
}

export type UpdateStatusBodyStatus =
  (typeof UpdateStatusBodyStatus)[keyof typeof UpdateStatusBodyStatus];

export const UpdateStatusBodyStatus = {
  new: "new",
  reviewed: "reviewed",
  escalated: "escalated",
} as const;

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}

export interface Alert {
  id: number;
  created_at: string;
  district?: string | null;
  signal_count?: number | null;
  time_window_hours?: number | null;
  risk_level?: string | null;
  cluster_summary?: string | null;
  status?: string | null;
}

export type UpdateAlertStatusBodyStatus =
  (typeof UpdateAlertStatusBodyStatus)[keyof typeof UpdateAlertStatusBodyStatus];

export const UpdateAlertStatusBodyStatus = {
  active: "active",
  acknowledged: "acknowledged",
  escalated: "escalated",
} as const;

export interface UpdateAlertStatusBody {
  status: UpdateAlertStatusBodyStatus;
}

export interface TimeseriesPoint {
  hour: string;
  total: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
}

export interface DrugAnalytic {
  drug_name: string;
  count: number;
}

export interface HospitalAnalytic {
  hospital_name: string;
  count: number;
}

export interface DistrictAnalytic {
  district: string;
  total: number;
  critical: number;
  high: number;
  medium: number;
  low: number;
  location_type?: string | null;
}

export type GetSignalsParams = {
  risk_level?: GetSignalsRiskLevel;
  category?: GetSignalsCategory;
  source_type?: GetSignalsSourceType;
  location_district?: string;
  location_type?: GetSignalsLocationType;
  from_date?: string;
  to_date?: string;
  search?: string;
  limit?: number;
  offset?: number;
};

export type GetSignalsRiskLevel =
  (typeof GetSignalsRiskLevel)[keyof typeof GetSignalsRiskLevel];

export const GetSignalsRiskLevel = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type GetSignalsCategory =
  (typeof GetSignalsCategory)[keyof typeof GetSignalsCategory];

export const GetSignalsCategory = {
  adr: "adr",
  hospital_issue: "hospital_issue",
  outbreak_signal: "outbreak_signal",
  treatment_complication: "treatment_complication",
} as const;

export type GetSignalsSourceType =
  (typeof GetSignalsSourceType)[keyof typeof GetSignalsSourceType];

export const GetSignalsSourceType = {
  social_media: "social_media",
  whatsapp: "whatsapp",
  hospital_form: "hospital_form",
  field_worker: "field_worker",
} as const;

export type GetSignalsLocationType =
  (typeof GetSignalsLocationType)[keyof typeof GetSignalsLocationType];

export const GetSignalsLocationType = {
  urban: "urban",
  rural: "rural",
} as const;