# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedWatch AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /signals:
    get:
      operationId: getSignals
      tags: [signals]
      summary: Get all signals
      parameters:
        - name: risk_level
          in: query
          schema:
            type: string
            enum: [critical, high, medium, low]
        - name: category
          in: query
          schema:
            type: string
            enum: [adr, hospital_issue, outbreak_signal, treatment_complication]
        - name: source_type
          in: query
          schema:
            type: string
            enum: [social_media, whatsapp, hospital_form, field_worker]
        - name: location_district
          in: query
          schema:
            type: string
        - name: location_type
          in: query
          schema:
            type: string
            enum: [urban, rural]
        - name: from_date
          in: query
          schema:
            type: string
        - name: to_date
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignalsListResponse"
...
  /signals/analyze:
    post:
      operationId: analyzeSignal
      tags: [signals]
      summary: Submit and analyze a signal via AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeSignalBody"
      responses:
        "200":
          description: Analyzed signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
...
  /signals/{id}:
    get:
      operationId: getSignal
      tags: [signals]
      summary: Get a single signal by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Signal detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /signals/{id}/status:
    patch:
      operationId: updateSignalStatus
      tags: [signals]
      summary: Update signal status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated signal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Signal"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Signal:
      type: object
      properties:
        id:
          type: integer
        created_at:
          type: string
        incident_date:
          type: string
          nullable: true
        source_type:
          type: string
          nullable: true
        location_district:
          type: string
          nullable: true
        location_type:
          type: string
          nullable: true
        raw_text:
          type: string
          nullable: true
        reporter_type:
          type: string
          nullable: true
        drug_name:
          type: string
          nullable: true
        hospital_name:
          type: string
          nullable: true
        risk_level:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        extracted_entities:
          type: string
          nullable: true
        recommended_action:
          type: string
          nullable: true
        status:
          type: string
          nullable: true
        nlp_summary:
          type: string
          nullable: true
      required:
        - id
        - created_at
    SignalsListResponse:
      type: object
      properties:
        signals:
          type: array
          items:
            $ref: "#/components/schemas/Signal"
        total:
          type: integer
      required:
        - signals
        - total
...
    AnalyzeSignalBody:
      type: object
      properties:
        raw_text:
          type: string
        source_type:
          type: string
        location_district:
          type: string
        location_type:
          type: string
        reporter_type:
          type: string
        drug_name:
          type: string
          nullable: true
        hospital_name:
          type: string
          nullable: true
        incident_date:
          type: string
          nullable: true
      required:
        - raw_text
    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, reviewed, escalated]
      required:
        - status
...