# lib/api-zod/src/generated/types/analyzeSignalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedWatch AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AnalyzeSignalBody {
  raw_text: string;
  source_type?: string;
  location_district?: string;
  location_type?: string;
  reporter_type?: string;
  drug_name?: string | null;
  hospital_name?: string | null;
  incident_date?: string | null;
}