# lib/api-zod/src/generated/types/updateStatusBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MedWatch AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateStatusBodyStatus } from "./updateStatusBodyStatus";

export interface UpdateStatusBody {
  status: UpdateStatusBodyStatus;
}